Database file split up in multiple files to make it more organized
